refactor: implement search functionality in new SearchRepository and SearchRepositoryImpl
